+++ ChangeLog.md
@@ -1,5 +1,10 @@
 
 ## Changes ##
+ - ENH: Add new Semantic Type - STREET_ADDRESS3_EN (Third line of an address)
+ - ENH: Continue to improve address detection
+
+### 12.4.0
+ - ENH: Bump google phone number library
 
 ### 12.3.3
  - ENH: Bump google phone number library